Remove current (incomplete) period from vote count  graph
