+++ src/lib/misc.ts
@@ -655,6 +655,16 @@ export function getCurrentGameId() {
     return null;
 }
 
+// x is a  date, y is data for that date
+export function dropCurrentPeriod(data: { x: string; y: number | null }[]) {
+    const newData = [...data];
+    if (newData.length > 0) {
+        const lastIndex = newData.length - 1;
+        newData[lastIndex] = { ...newData[lastIndex], y: null };
+    }
+    return newData;
+}
+
 /* This code is hacked together from
         https://github.com/simov/slugify/blob/master/slugify.js and
         https://github.com/dodo/node-slug/blob/master/slug.js
