+++ src/views/User/VoteActivityGraph.tsx
@@ -20,6 +20,7 @@ import React, { useEffect, useState } from "react";
 import { get } from "requests";
 import * as data from "data";
 import { ResponsiveLine } from "@nivo/line";
+import { dropCurrentPeriod } from "misc";
 
 interface VoteCountPerDay {
     date: string; // assuming the date is returned as a string, e.g., "2024-03-17"
@@ -67,7 +68,7 @@ const VoteActivityGraph = ({ vote_data }: VoteActivityGraphProps) => {
         return [
             {
                 id: "votes",
-                data,
+                data: dropCurrentPeriod(data),
             },
         ];
     }, [vote_data]);
